первое разбиение на модули(utill, data)
