+++ js/util.js
@@ -1,0 +1,23 @@
+let userID = 0;
+let commentID = 0;
+
+// eslint-disable-next-line no-undef, no-unused-vars
+const getUsersID = () => ++userID;
+
+// eslint-disable-next-line no-undef, no-unused-vars
+const getCommentsID = () => ++commentID;
+
+const getRandomPositiveInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// eslint-disable-next-line no-unused-vars
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+// eslint-disable-next-line no-unused-vars
+const checkStringLength = (message, maxLength) => message.length <= maxLength;
+
+export {getRandomArrayElement, getCommentsID, getUsersID};
